fix: change REAL to DOUBLE PRECISION in opportunities and spend migrations
PostgreSQL REAL is float4 (f32); sqlx maps it to Rust f32, not f64.
The Rust models use f64 for amount/amount_usd, so the column type must
be DOUBLE PRECISION (float8) to avoid a decode type mismatch.

+++ opportunities-service/migrations/0001_create_opportunities.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS opportunities (
     account_id TEXT    NOT NULL,
     name       TEXT    NOT NULL,
     stage      TEXT    NOT NULL DEFAULT 'qualification',
-    amount     REAL    NOT NULL DEFAULT 0.0,
+    amount     DOUBLE PRECISION NOT NULL DEFAULT 0.0,
     close_date TEXT,
     created_at TEXT    NOT NULL DEFAULT (to_char(timezone('UTC', now()), 'YYYY-MM-DD"T"HH24:MI:SS"Z"')),
     updated_at TEXT    NOT NULL DEFAULT (to_char(timezone('UTC', now()), 'YYYY-MM-DD"T"HH24:MI:SS"Z"'))
